Make compareBoundaryPoints match browsers
First: we check for detached ranges before we check for the validity of the
first parameter.  This matches all browsers tested (Firefox 12.0a1, Chrome 17
dev, Opera Next 12.00 alpha, and I think IE9).

Second: we throw NotSupportedError instead of IndexSizeError for incorrect
values of the first parameter.  The only browser to throw a standard exception
type here at all is Opera, which throws NotSupportedError, so let's reward them
by making them right.  It makes more sense anyway, since the first parameter
isn't an index.

Third: START_TO_END compares the start of the *argument* to the end of the
*context object*, and END_TO_START is vice versa, in Firefox 12.0a1, Chrome 17
dev, and Opera Next 12.00 alpha.  (I think IE9 too, but couldn't test.)  Test
case:

data:text/html,<!doctype html>
<script>
var range = document.createRange();
range.setEnd(document.head, 0);
alert(document.createRange().compareBoundaryPoints(
  Range.START_TO_END, range));
</script>

This alerts 0.  But the start of document.createRange() is not the same as the
end of range -- it's the start of range that's the same as the end of
document.createRange().