`document.importNode` and `document.adoptNode` should error when acting upon ShadowRoots
The `importNode` algorithm relies on a `clone` of `node` however, the Shadow
DOM spec stipulates that cloning a `ShadowRoot` should raise a
`DataCloneError`. Consequently, `document.importNode` on a ShadowRoot should
be unsupported.

Additionally, it doesn't make sense for a document to `adopt` a `ShadowRoot` -
the `ShadowRoot` should always be attached to a shadow host.  In this
instance a `HierarchyRequestError` seems to make sense.