Shadow: add shadow root
This is basically the introduction of ShadowRoot objects as a node in
the DOM Standard. The plan is to port over the features over time,
while trying to make sure the new features and terminology remain
consistent and coherent with the rest of the standard.

I’ll try to prefix ShadowRoot-object-related commits with “Shadow”.