Shadow: define composedPath()
Fixes #242. Also reverts
https://github.com/w3c/webcomponents/issues/432 as it appears to no
longer be needed. If we ever have an event path that starts with a
non-node and ends with a node we'll have to add that complexity back.

(The reason it is named composedPath() is because deepPath() and
various other names were no longer viable after Chrome's initial
experiment. See the various issues.)