Add getRootNode()
Unfortunately a variety of good short names is incompatible with the
web.

Furthermore, the argument was made that a method better exposes the
fact that this is not always a O(1) operation and can be used to return
the shadow-including root when given an argument.

Fixes #241.

(Although 42a45eb0d1af13e510620b6b29696997f0c7aa3d indicated that no
consensus could be found for an alternative to
Node.prototype.rootNode, removing that property fast tracked
consensus for this method.)