Define special dispatch behavior for click events
User agents have always had special behavior for synthetic click
events. This is discussed to great extent in
https://www.w3.org/Bugs/Public/show_bug.cgi?id=12230.

This change introduces the hooks to no longer make that magical, but an
understood and integral part of the processing model, ugly as it is.

This changes requires corresponding changes and cleanup in HTML. See
https://github.com/whatwg/html/issues/1394 for that discussion.

Fixes #325.