Reinstate active flag for iterators
This reverts ccf51e85, and fixes
https://www.w3.org/Bugs/Public/show_bug.cgi?id=25412.  It brings the
spec back in line with Firefox instead of other browsers.  Boris argues
convincingly that the behavior without the active flag is much harder to
spec and test properly, and nobody has a use-case for allowing recursion
here, so it's easier all around to converge on Firefox behavior and just
ban recursion.  He doesn't want Firefox to allow recursion unless
someone writes good tests for it, which nobody wants to.  Other UAs were
asked for comment on the bug and didn't support or oppose.