Clarify insert-node's record's previousSibling
This adds a previousSibling variable to the insert node algorithm in
order to clarify that the previousSibling property of the mutation
record generated by the insert node algorithm should be determined
before inserting nodes. If implemented as described in the original
text, the property would be determined after insertion, and will
therefore always be the last of nodes.