Make slots do nothing outside shadow trees
A slot's children are no longer considered as the slot's fallback contents if
the slot is not in a shadow tree. Such a slot no longer signals a slot change
even if its children are updated.

This patch also removes a concept of suppress signaling flag for a slot change
entirely. A slot change is always signaled when a slot's assigned nodes are
changed.

Fixes #447.