Make the DOMTokenList creation steps use "get an attribute value"
This makes it so that if the `DOMTokenList`'s associated element has an attribute that is the associated attribute's local name, at creation time, it will use the "get an attribute value" steps, which will always return a string. Previously, the this used the "get an attribute by namespace and local name" steps, whuch would return an `Attr` node if the associated element had an attribute that is the associated attribute's local name. Fixes #666.