Export concept of collapsed
The Selection API makes extensive reference to the "collapsed" concept. It's currently explicitly linked through an absolute URL. Alternatively, it might be nice to export it in such a way that it can be used via [=range/collapsed=] - for whatever  reason, that form is not being exposed to BikeShed/Shepherd's data.