Limit AbortSignal.timeout to Window and Worker
Worklets are designed with minimal event loops such that they need only
run a single task and corresponding microtasks, making APIs that
schedule additional tasks unsuitable for worklets. As such, we restrict
the availability of AbortSignal.timeout to Window and Worker since the
API schedules a task to implement the timeout.

Closes #1054.