Typo occurance -> occurrence in section id
This commit fixes a small typo in the id to the section https://dom.spec.whatwg.org/#action-versus-occurance
Since this id is probably used in many places we kept a `<span>` with the id to preserve the previous link.
I'm not too used to bikeshed, but for whatever reasons it was requesting the id to be hardcoded, instead of deriving it from the title's content.