Define slotchange event
Following discussions in https://github.com/whatwg/html/issues/8340, the events
index in the HTML spec will be restricted to event types that gets fired within
the HTML spec. The `slotchange` event will be dropped as a result.

This turns the mention of the `slotchange` event into a proper event type
definition, so that other specs (including the HTML spec) can reference it.