Enforce same parameters for attachShadow on declarative shadow root
Per the discussion at https://github.com/whatwg/dom/issues/1235, this PR changes the behavior when calling `attachShadow()` on a node with an existing declarative shadow root. Now, if the parameters do not match between the arguments to `attachShadow()` and the existing root, throw an exception. Prior behavior was to silently return the declarative root as-is, with mismatched parameters.