Enforce mode matching for attachShadow() on declarative shadow root
This changes the behavior when calling `attachShadow()` on a node with an existing declarative shadow root. Now, if mode does not match between the arguments to `attachShadow()` and the existing root, throw an exception. Prior behavior was to silently return the declarative root as-is, with mismatched parameters.

Additionally, remove the note as it was incorrect.

See also: https://github.com/whatwg/html/pull/10069.

Fixes #1235.

Co-authored-by: Anne van Kesteren <annevk@annevk.nl>