Editorial: Update to use new "report an exception" algorithm in HTML
It's necessary to specify which global these are reported to. This seems
to be the event listener callback's realm's global in the event listener
case, and the custom element constructor's realm, based on a combination
of consistency with the related sites in HTML and browser behavior --
though browser behavior for custom elements isn't consistent in the
(unusual) case of custom elements across realms.

Part of whatwg/html#10516.