Remove quadratic behavior from clone a node
Strictly speaking this should be an editorial change, but as it's fairly significant not marking it as such. This improves integration with HTML as well.

As part of this change we move the shadow tree behavior under creating the element clone as that is a more logical fit.

Unfortunately subtree has to be passed everywhere as HTML needs it for its template element logic. That's also why clone a node's children is exported.

Corresponding HTML PR: SOON.

Fixes #1219. Closes #1220.