Remove quadratic behavior from clone a node
Strictly speaking this should be an editorial change, but as it's fairly significant not marking it as such. This improves integration with HTML as well.

All the arguments are exported as HTML needs them as well. And they are named parameters now for clarity.

Corresponding HTML PR: https://github.com/whatwg/html/pull/10859.

Fixes #1219. Closes #1220.